Not used by multiple presenters anymore
