+++ app/presenters/capital_repayment_holiday_loan_change.rb
@@ -1,8 +1,14 @@
 class CapitalRepaymentHolidayLoanChange < LoanChangePresenter
-  include CapitalRepaymentHolidayFields
+  delegate :initial_capital_repayment_holiday, :initial_capital_repayment_holiday=, to: :premium_schedule
+
+  attr_accessible  :initial_capital_repayment_holiday
 
   validate :validate_capital_repayment_holiday
   before_save :update_loan_change
+
+  def includes_capital_repayment_fields?
+    true
+  end
 
   private
     def update_loan_change
